Rename reducer.ts to utils
